Split large files into seperate files
