+++ mlir/lib/Dialect/MIOpen/Tuning/CMakeLists.txt
@@ -1,8 +1,13 @@
 add_mlir_dialect_library(MLIRMIOpenTuning
   SqliteDb.cpp
   GridwiseGemmParams.cpp
-  conv_hip_implicit_gemm.cpp
-  conv_hip_implicit_gemm_xdl.cpp
+  solver/conv_hip_implicit_gemm_fwd_v4r4.cpp
+  solver/conv_hip_implicit_gemm_wrw_v4r4.cpp
+  solver/conv_hip_implicit_gemm_bwd_v1r1.cpp
+  solver/conv_hip_implicit_gemm_fwd_v4r4_xdlops.cpp
+  solver/conv_hip_implicit_gemm_wrw_v4r4_xdlops.cpp
+  solver/conv_hip_implicit_gemm_bwd_v4r1_xdlops.cpp
+
 
   ADDITIONAL_HEADER_DIRS
   ${MLIR_MAIN_INCLUDE_DIR}/mlir/Dialect/MIOpen/Tuning
